renamed client.py to test_client for datanode
